requirements.txt
# Core Web Framework
Django==3.0.8

# Deep Learning & Machine Learning
tensorflow==2.10.0
keras==2.10.0
numpy==1.23.5
pandas==1.5.3

# Image Processing
Pillow==9.5.0

# Database
# SQLite3 comes with Python, no separate installation needed

# Optional but recommended
matplotlib==3.7.1
scikit-learn==1.2.2

# Development Tools (optional)
ipython==8.12.0
jupyter==1.0.0
notebook==6.5.4

# Additional utilities
h5py==3.8.0
